adding to readme and amended makefile

+++ Makefile
@@ -3,7 +3,7 @@
 default: build container
 
 .PHONY: gobuild
-build:
+gobuild:
 	go get github.com/lukemgriffith/autodns
 	env CGO_ENABLED=0 go build -o autodns . 
 
@@ -15,14 +15,22 @@ container:
 
 .PHONY: deploy
 deploy:
-	kubectl apply -f manifest.json
+	#kubectl apply -f manifest.json
+	docker run -e 'AWS_ACCESS_KEY_ID=$(AWS_ACCESS_KEY_ID)' \
+		-e 'AWS_SECRET_ACCESS_KEY=$(AWS_SECRET_ACCESS_KEY)' \
+		-e 'zone_name=griffith.cloud' \
+		-e 'a_record=home' \
+		-e 'tfDir=/app/workspace' -d \
+		lukegriffith/autodns:latest
 
 .PHONY: clean
 clean:
-	for pod in `kubectl get pods | grep autodns | awk '{print $$1}'`; do kubectl delete pods $$pod; done;
 
-.PHONY: setup
-terraform:
-	curl -o /tmp/terraform.zip https://releases.hashicorp.com/terraform/0.11.13/terraform_0.11.13_linux_amd64.zip 
-	unzip /tmp/terraform.zip
+.PHONY: deps
+deps:
+	if [ ! -f ./terraform ]
+	then
+		curl -o /tmp/terraform.zip https://releases.hashicorp.com/terraform/0.11.13/terraform_0.11.13_linux_amd64.zip 
+		unzip /tmp/terraform.zip
+	fi
